# docs/swagger.yaml
basePath: /
definitions:
  models.Department:
    properties:
      color:
        type: string
      created_at:
        type: string
      employees:
        items:
          $ref: '#/definitions/models.Employee'
        type: array
      id:
        type: integer
      name:
        type: string
      shift_weeks:
        items:
          $ref: '#/definitions/models.ShiftWeek'
        type: array
      updated_at:
        type: string
    type: object
  models.Employee:
    properties:
      color:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ShiftWeek:
    properties:
      created_at:
        type: string
      department_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      notes:
        type: string
      start_date:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  responses.APIResponse:
    properties:
      data: {}
      error:
        example: Fehler bei der Verarbeitung
        type: string
      message:
        example: Operation erfolgreich
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Carsten Bröckert
  description: REST-API für die Verwaltung von Schichtplänen
  license:
    name: MIT
  title: Schichtplaner API
  version: "1.0"
paths:
  /api/v1/departments:
    get:
      consumes:
      - application/json
      description: Ruft alle Abteilungen mit zugehörigen Mitarbeitern und Schichtwochen
        ab
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Department'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Alle Abteilungen abrufen
      tags:
      - departments
    post:
      consumes:
      - application/json
      description: Erstellt eine neue Abteilung mit den angegebenen Daten
      parameters:
      - description: Abteilungsdaten
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Department'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Abteilung erstellen
      tags:
      - departments
  /api/v1/departments/{id}:
    delete:
      consumes:
      - application/json
      description: Löscht eine Abteilung und ihre Beziehungen
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Abteilung löschen
      tags:
      - departments
    get:
      consumes:
      - application/json
      description: Ruft eine spezifische Abteilung mit allen Details ab
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Department'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Einzelne Abteilung abrufen
      tags:
      - departments
    put:
      consumes:
      - application/json
      description: Aktualisiert eine bestehende Abteilung
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      - description: Aktualisierte Abteilungsdaten
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Department'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Abteilung aktualisieren
      tags:
      - departments
  /api/v1/departments/{id}/employees:
    get:
      consumes:
      - application/json
      description: Ruft alle Mitarbeiter einer spezifischen Abteilung ab
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Employee'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Mitarbeiter einer Abteilung abrufen
      tags:
      - departments
  /api/v1/departments/{id}/stats:
    get:
      consumes:
      - application/json
      description: Ruft statistische Daten einer Abteilung ab
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Abteilungsstatistiken abrufen
      tags:
      - departments
...
swagger: "2.0"